api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-generativeai
lunarcalendar
